app/requirements.txt
flask
pandas
tuya-connector-python
mysql.connector